# api/openapi.yaml
openapi: 3.0.3
info:
  title: Codificar Chamados API
  version: 1.0.0
  description: API de chamados usada pelo frontend do projeto.
servers:
  - url: http://localhost:8080/api

tags:
  - name: Tickets
    description: Operações de chamados
  - name: Assignees
    description: Operações de responsáveis

paths:
  /tickets:
    get:
      tags: [Tickets]
      summary: Lista chamados
      description: |
        Retorna os chamados ordenados por chamados ativos primeiro, prioridade e data de abertura.
        O filtro `search` hoje é aplicado apenas ao título.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/TicketPriority'
        - name: assignee_id
          in: query
          schema:
            type: integer
            format: int64
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de chamados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Tickets]
      summary: Cria chamado
      description: Se `assignee_id` for `0`, o backend faz atribuição automática.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketInput'
      responses:
        '201':
          description: Chamado criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: Obtém um chamado
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Chamado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Tickets]
      summary: Exclui chamado
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '204':
          description: Chamado excluído
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Tickets]
      summary: Atualiza chamado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdateInput'
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Chamado atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /assignees:
    get:
      tags: [Assignees]
      summary: Lista responsáveis
      responses:
        '200':
          description: Lista de responsáveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignee'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /assignees/workload:
    get:
      tags: [Assignees]
      summary: Carga por responsável
      responses:
        '200':
          description: Carga dos responsáveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssigneeWorkload'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    TicketId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    TicketStatus:
      type: string
      enum:
        - open
        - in_progress
        - resolved
        - closed

    TicketPriority:
      type: string
      enum:
        - low
        - medium
        - high

    Ticket:
      type: object
      required:
        - id
        - title
        - description
        - priority
        - status
        - assignee_id
        - opened_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        assignee_id:
          type: integer
          format: int64
          nullable: true
        opened_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    TicketInput:
      type: object
      required:
        - title
        - priority
        - assignee_id
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          default: ""
        priority:
          $ref: '#/components/schemas/TicketPriority'
        assignee_id:
          type: integer
          format: int64
          description: Use 0 to let the backend choose automatically.

    TicketUpdateInput:
      allOf:
        - $ref: '#/components/schemas/TicketInput'
        - type: object
          required:
            - status
          properties:
            status:
              $ref: '#/components/schemas/TicketStatus'
            auto_reassign:
              type: boolean
              default: false

    Assignee:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        last_assigned_at:
          type: string
          format: date-time
          nullable: true

    AssigneeWorkload:
      type: object
      required:
        - assignee_id
        - name
        - open_count
      properties:
        assignee_id:
          type: integer
          format: int64
        name:
          type: string
        open_count:
          type: integer

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string